chore: org files and add logging
