+++ apps/trading-e2e/src/support/capsule/faucet-asset.ts
@@ -1,10 +1,19 @@
-export async function faucetAsset(assetId: string, vegaPubKey: string) {
+import { createLog } from './logging';
+
+const log = createLog('faucet-asset');
+
+export async function faucetAsset(
+  asset: string,
+  party: string,
+  amount = '10000'
+) {
+  log(`sending ${amount} ${asset} to ${party}`);
   const res = await fetch('http://localhost:1790/api/v1/mint', {
     method: 'post',
     body: JSON.stringify({
-      amount: '10000',
-      asset: assetId,
-      party: vegaPubKey,
+      amount,
+      asset,
+      party,
     }),
   });
   const json = await res.json();
