+++ apps/trading-e2e/src/support/capsule/logging.ts
@@ -1,0 +1,5 @@
+export function createLog(name: string) {
+  return (message: string) => {
+    console.log(`[${name}]: ${message}`);
+  };
+}
