+++ apps/trading-e2e/src/support/utils.ts
@@ -1,9 +1,11 @@
 import { ethers } from 'ethers';
 import sha3 from 'js-sha3';
 import type { Transaction } from '@vegaprotocol/wallet';
+import BigNumber from 'bignumber.js';
 
 /**
- * Base64 encode a transaction object
+ * copy of encodeTransaction in libs/wallet/src/utils.ts
+ * to avoid pulling in any jsx files which will cypress is not set up to compile
  */
 export const encodeTransaction = (tx: Transaction): string => {
   return ethers.utils.base64.encode(
@@ -12,8 +14,18 @@ export const encodeTransaction = (tx: Transaction): string => {
 };
 
 /**
- * Convert a Vega transaction signature to an object ID
+ * copy of determineId in libs/wallet/src/utils.ts
+ * to avoid pulling in any jsx files which will cypress is not set up to compile
  */
 export function determineId(sig: string) {
   return sha3.sha3_256(ethers.utils.arrayify('0x' + sig));
+}
+
+/**
+ * copy of removeDecimal from libs/react-helpers/src/lib/format/number.tsx
+ * to avoid pulling in any jsx files which will cypress is not set up to compile
+ */
+export function removeDecimal(value: string, decimals: number): string {
+  if (!decimals) return value;
+  return new BigNumber(value || 0).times(Math.pow(10, decimals)).toFixed(0);
 }
